Use library for one-off events + show drawer when first used
Until the user has opened the navigation drawer himself, open it when the app starts.
This used to be a Google recommendation, but with the transition it was lost. This doesn't make it any less valid though!

See: http://web.archive.org/web/20150315013505/https://developer.android.com/design/patterns/navigation-drawer.html