Move to new built-in RV adapters (WIP)
Create parallel adapter implementations using the new adapters from the
support library as a base. Deprecate the old classes.

This currently has some bugs when the new adapters are used with empty
views. The empty view is never removed, since the callback order has
changed.

This issue is tracked at i/74386483, to get feedback from Google if this
is intended or not. Their answer will determine how this will eventually
be fixed.