Use the correct opening hours
Because of caching and the API not being updated instantly, sometimes
the first opening hours in the list are the ones of yesterday. This
commit ensures we always use the opening hours for today.

The code can now handle all responses from the API; for example, if the
API skips some days.