Introduce HackedViewPager in attempt to fix crashes
A search on Google will surface multiple reports of slightly different,
yet similar issues, all relating to intercepting touch events.
One of the most recommended fixes is to just catch and ignore the
exceptions, which we do now.

The errors are still logged to give us an idea of the error rate, and
to be able to check for errors that were not supposed to be caught.