Remove the app logo
This (very old) logo is no longer used. The android:logo attribute can
be displayed on the ActionBar if the appropriate flags are set; I don't
think Hydra ever did this and if it did, it was a long time ago.

Should we ever need to add a logo again, we will have to create a new
one anyway.