Ignore exception when starting warm-up service for Custom Tabs
Some more details (see also the source code comment). The reason why
this exception occurs is probably as follows. On Android 8 and above,
an app can no longer start services when it is in the background.
However, to warm the Custom Tabs, a service has to be started. There is
a very small but real possibility that the user opens an activity with
custom tabs and navigates away from the app after the connection has
been initialized, but before the actual warm-up service has started.
This theory is supported by the crash logs which only indicate
Android 8+, and various only bug reports.

See also e.g. https://github.com/openid/AppAuth-Android/issues/425