Don't add favourite product multiple times
The cart is saved when the activity is closed and restored when opening
the activity again. This produces the unintuitive behaviour that using
the favourite button again results in adding the product again, even if
it was still in the cart from a previous time.

This prevents that by not adding the product if it already present.